# [TRANSFORM] use precompiled JavaScript for faster transformation