--- a/extensions/transform/org.eclipse.smarthome.transform.javascript/src/main/java/org/eclipse/smarthome/transform/javascript/internal/JavaScriptTransformationService.java
+++ b/extensions/transform/org.eclipse.smarthome.transform.javascript/src/main/java/org/eclipse/smarthome/transform/javascript/internal/JavaScriptTransformationService.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2014,2018 Contributors to the Eclipse Foundation
+ * Copyright (c) 2014,2019 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -12,23 +12,18 @@
  */
 package org.eclipse.smarthome.transform.javascript.internal;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.InputStreamReader;
-import java.io.Reader;
+import java.util.Objects;
 
-import javax.script.ScriptEngine;
-import javax.script.ScriptEngineManager;
+import javax.script.Bindings;
+import javax.script.CompiledScript;
 import javax.script.ScriptException;
 
-import org.apache.commons.io.IOUtils;
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.eclipse.smarthome.config.core.ConfigConstants;
 import org.eclipse.smarthome.core.transform.TransformationException;
 import org.eclipse.smarthome.core.transform.TransformationService;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,10 +32,23 @@ import org.slf4j.LoggerFactory;
  * input by Java Script.
  *
  * @author Pauli Anttila
+ * @author thomask
  */
 @NonNullByDefault
 @Component(immediate = true, property = { "smarthome.transform=JS" })
 public class JavaScriptTransformationService implements TransformationService {
+
+    private Logger logger = LoggerFactory.getLogger(JavaScriptTransformationService.class);
+    private @NonNullByDefault({}) JavaScriptEngineManager manager;
+
+    @Reference
+    public void setJavaScriptEngineManager(JavaScriptEngineManager manager) {
+        this.manager = manager;
+    }
+
+    public void unsetJavaScriptEngineManager(JavaScriptEngineManager manager) {
+        this.manager = null;
+    }
 
     /**
      * Transforms the input <code>source</code> by Java Script. It expects the
@@ -49,49 +57,31 @@ public class JavaScriptTransformationService implements TransformationService {
      * transformations one should use subfolders.
      *
      * @param filename the name of the file which contains the Java script
-     *            transformation rule. Transformation service inject input
-     *            (source) to 'input' variable.
-     * @param source the input to transform
+     *                     transformation rule. Transformation service inject input
+     *                     (source) to 'input' variable.
+     * @param source   the input to transform
      */
     @Override
     public @Nullable String transform(String filename, String source) throws TransformationException {
-        Logger logger = LoggerFactory.getLogger(JavaScriptTransformationService.class);
-        if (filename == null || source == null) {
-            throw new TransformationException("the given parameters 'filename' and 'source' must not be null");
-        }
+        Objects.requireNonNull(filename, "the given parameter 'filename' must not be null");
+        Objects.requireNonNull(source, "the given parameter 'source' must not be null");
 
-        logger.debug("about to transform '{}' by the Java Script '{}'", source, filename);
+        final long startTime = System.currentTimeMillis();
+        logger.debug("about to transform '{}' by the JavaScript '{}'", source, filename);
 
-        Reader reader;
-
-        try {
-            String path = ConfigConstants.getConfigFolder() + File.separator
-                    + TransformationService.TRANSFORM_FOLDER_NAME + File.separator + filename;
-            reader = new InputStreamReader(new FileInputStream(path));
-        } catch (FileNotFoundException e) {
-            throw new TransformationException("An error occurred while loading script.", e);
-        }
-
-        ScriptEngineManager manager = new ScriptEngineManager();
-
-        ScriptEngine engine = manager.getEngineByName("javascript");
-        engine.put("input", source);
-
-        Object result = null;
-
-        long startTime = System.currentTimeMillis();
+        String result = "";
 
         try {
-            result = engine.eval(reader);
+            final CompiledScript cScript = manager.getScript(filename);
+            final Bindings bindings = cScript.getEngine().createBindings();
+            bindings.put("input", source);
+            result = String.valueOf(cScript.eval(bindings));
+            return result;
         } catch (ScriptException e) {
-            throw new TransformationException("An error occurred while executing script.", e);
+            throw new TransformationException("An error occurred while executing script. " + e.getMessage(), e);
         } finally {
-            IOUtils.closeQuietly(reader);
+            logger.trace("JavaScript execution elapsed {} ms. Result: {}", System.currentTimeMillis() - startTime,
+                    result);
         }
-
-        logger.trace("JavaScript execution elapsed {} ms", System.currentTimeMillis() - startTime);
-
-        return String.valueOf(result);
     }
-
 }