On an ARM machine transformation by JavaScript [takes ~15 seconds each time](https://community.openhab.org/t/js-transformation-takes-10-15-seconds-each-call/61566). To speed up this a little bit, each JavaScript file will be pre compiled and cached. This reduces the time of transformation to ~500ms. First call of each JavaScript still takes ~15 seconds. each further call needs then ~500ms.
Transformation directory will be observed for file changes and the cache will be refreshed.

This implementation is similar to [org.eclipse.smarthome.automation.module.script](https://github.com/eclipse/smarthome/tree/master/bundles/automation/org.eclipse.smarthome.automation.module.script)