@kaikreuzer 2019-01-16
@THKDev Please see https://github.com/eclipse/smarthome/blob/master/CONTRIBUTING.md#legal-considerations how to sign the ECA - thanks!